fix(process-metrics): build on non-Linux platforms
- Gate procfs dependency to cfg(target_os = "linux") so its build script
  no longer fails on macOS
- Split into a portable ProcessMetrics (memory, cpu_count, uptime via
  sysinfo) and a Linux-only LinuxMetrics extension (/proc, cgroup, PSI)
- Keep the only platform cfg in lib.rs; non-Linux gets an inline no-op ext
